Enabled cleanup for streaming channels.
The previous implementation closed the message and error streams before Closing
the underlying stream transport. When the underlying transport is canceled, it
returns with an error, the go func tries to send the error on the error channel,
but causes a panic because the error channel had already been closed.

Instead, we let the go func close the message and error channels when the
underlying transport returns with its error. Closing the streamer now only
involves Close'ing the underlying transport. The error message is enqueued to
the error channel, after which the go func returns and both channels close.